feat: [mig-runtime-api-held-destination-cleared-deletion-batches-005] - Delete read_save-mapped catch-all Tests
